+++ opentelemetry-exporter-gcp-logging/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+Added support for when `bytes` or `list['bytes']` is in `LogRecord.body` and
+body is of type Mapping. Update opentelemetry-api/sdk dependencies to 1.3.
+
 ## Version 1.9.0a0
 
 Released 2025-02-03