+++ opentelemetry-exporter-gcp-logging/CHANGELOG.md
@@ -7,6 +7,9 @@
 - Do not call `logging.warning` when `LogRecord.body` is of None type, instead leave `LogEntry.payload` empty.
 - Update opentelemetry-api/sdk dependencies to 1.3.
 
+The suffix part of `LogEntry.log_name` will be the `LogRecord.event_name` when
+that is present and the `gcp.log_name` attribute is not.
+
 ## Version 1.9.0a0
 
 Released 2025-02-03